🧑‍💻 Create a `HasLaracord` trait for accessing the Bot methods
